+++ Targets/STM32F4xx/STM32F4_CAN.cpp
@@ -323,7 +323,7 @@ struct CanState {
 
     bool isOpened;
 
-    uint32_t initializeCount;
+    uint16_t initializeCount;
 };
 
 static const STM32F4_Gpio_Pin canTxPins[] = STM32F4_CAN_TX_PINS;
